requirements.txt
google-adk
streamlit
pydantic
mistralai
qdrant-client
pandas
python-dotenv
google-genai